style: #6 로그인 화면 스타일 분리
