+++ package.json
@@ -44,8 +44,8 @@
     "@react-native-community/eslint-config": "^2.0.0",
     "@tsconfig/react-native": "^2.0.2",
     "@types/jest": "^26.0.23",
-    "@types/react": "^18.0.25",
-    "@types/react-native": "^0.70.7",
+    "@types/react": "^18.0.26",
+    "@types/react-native": "^0.70.8",
     "@types/react-native-material-dropdown": "^0.11.5",
     "@types/react-native-modal-dropdown": "^1.0.0",
     "@types/react-native-vector-icons": "^6.4.12",
@@ -57,7 +57,7 @@
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "0.72.3",
     "react-test-renderer": "18.1.0",
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.4"
   },
   "jest": {
     "preset": "react-native",
