+++ src/components/emailLogin/IDInputbox.tsx
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
-import {TextInput, StyleSheet} from 'react-native';
+import {TextInput} from 'react-native';
 import styles from '../../styles/emailLogin/IDInputbox';
 
 class InputboxID extends Component {
-    render() {
+    render () {
         return (
             <TextInput
                 style = {styles.inputStyle}
