+++ src/components/emailLogin/PWInputbox.tsx
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
-import {TextInput, StyleSheet} from 'react-native';
+import {TextInput} from 'react-native';
 import styles from '../../styles/emailLogin/PWInputbox';
 
 class PWInputbox extends Component {
-    render(){
+    render () {
         return (
             <TextInput
                 style = {styles.inputStyle}
