+++ src/styles/login/EmailLoginButton.js
@@ -1,0 +1,21 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    buttonStyle: {
+        marginTop: 10,
+        marginLeft: 20,
+        marginRight: 20,
+        padding: 12,
+        backgroundColor: '#2BC63B',
+        borderRadius: 8
+    },
+
+    buttonTitle: {
+        textAlign: 'center',
+        color: 'white', 
+        fontSize: 15,
+        fontFamily: 'GmarketSansTTFMedium'
+    }
+});
+
+export default styles;
