+++ src/styles/login/ForgotButton.js
@@ -1,0 +1,21 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    buttonStyle: {
+        marginTop: 10,
+        marginLeft: 20,
+        marginRight: 20,
+        padding: 8,
+        alignItems: 'center',
+        borderRadius: 8
+    },
+
+    buttonTitle: {
+        textAlign: 'center',
+        color: 'green', 
+        fontSize: 15,
+        fontFamily: 'GmarketSansTTFLight'
+    }
+});
+
+export default styles;
