+++ src/styles/login/GoogleLoginButton.js
@@ -1,0 +1,23 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    buttonStyle: {
+        marginTop: 10,
+        marginLeft: 20,
+        marginRight: 20,
+        padding: 12,
+        backgroundColor: 'white',
+        borderRadius: 8,
+        borderWidth: 1, 
+        borderColor: 'black'
+    },
+
+    buttonTitle: {
+        textAlign: 'center',
+        color: 'black', 
+        fontSize: 15,
+        fontFamily: 'GmarketSansTTFMedium'
+    }
+});
+
+export default styles;
